Get flight of individuals (BSO only)

Get the number of individuals that are flying.

Usage

bso_get_flight_rst(pheno, generation, flight = 1, dates = prop_dates(pheno))

bso_get_flight_df(
  pheno,
  generation,
  stations = prop_stations(pheno),
  flight = 1,
  dates = prop_dates(pheno)
)

Arguments

pheno

A BSO phenology (see bso_phenology())

generation

Generation of interest. For sister broods, 0.5 should be added.

flight

Specifies which flight of the respective generation should be returned. Can be 1 (first flight) or 2 (second flight).

dates

Select dates that should be present in the output.

stations

Pass a character vector to choose stations assigned to pheno by their names, or pass different stations. See stations_create() for details.

Value

bso_get_flight_rst: a multi-layer SpatRaster.

bso_get_flight_df: a data frame.

Examples

# \donttest{
# This may take a few minutes...

# calculate phenology
p <- bso_phenology('bso', barrks_data('stations'), .quiet = TRUE)

# get the number of individuals of the hibernating generation on their first flight
bso_get_flight_df(p, 0, 'Freiburg', flight = 1)
#>      station       date doy   flight
#> 1   Freiburg 2011-01-01   1        0
#> 2   Freiburg 2011-01-02   2        0
#> 3   Freiburg 2011-01-03   3        0
#> 4   Freiburg 2011-01-04   4        0
#> 5   Freiburg 2011-01-05   5        0
#> 6   Freiburg 2011-01-06   6        0
#> 7   Freiburg 2011-01-07   7        0
#> 8   Freiburg 2011-01-08   8        0
#> 9   Freiburg 2011-01-09   9        0
#> 10  Freiburg 2011-01-10  10        0
#> 11  Freiburg 2011-01-11  11        0
#> 12  Freiburg 2011-01-12  12        0
#> 13  Freiburg 2011-01-13  13        0
#> 14  Freiburg 2011-01-14  14        0
#> 15  Freiburg 2011-01-15  15        0
#> 16  Freiburg 2011-01-16  16        0
#> 17  Freiburg 2011-01-17  17        0
#> 18  Freiburg 2011-01-18  18        0
#> 19  Freiburg 2011-01-19  19        0
#> 20  Freiburg 2011-01-20  20        0
#> 21  Freiburg 2011-01-21  21        0
#> 22  Freiburg 2011-01-22  22        0
#> 23  Freiburg 2011-01-23  23        0
#> 24  Freiburg 2011-01-24  24        0
#> 25  Freiburg 2011-01-25  25        0
#> 26  Freiburg 2011-01-26  26        0
#> 27  Freiburg 2011-01-27  27        0
#> 28  Freiburg 2011-01-28  28        0
#> 29  Freiburg 2011-01-29  29        0
#> 30  Freiburg 2011-01-30  30        0
#> 31  Freiburg 2011-01-31  31        0
#> 32  Freiburg 2011-02-01  32        0
#> 33  Freiburg 2011-02-02  33        0
#> 34  Freiburg 2011-02-03  34        0
#> 35  Freiburg 2011-02-04  35        0
#> 36  Freiburg 2011-02-05  36        0
#> 37  Freiburg 2011-02-06  37        0
#> 38  Freiburg 2011-02-07  38        0
#> 39  Freiburg 2011-02-08  39        0
#> 40  Freiburg 2011-02-09  40        0
#> 41  Freiburg 2011-02-10  41        0
#> 42  Freiburg 2011-02-11  42        0
#> 43  Freiburg 2011-02-12  43        0
#> 44  Freiburg 2011-02-13  44        0
#> 45  Freiburg 2011-02-14  45        0
#> 46  Freiburg 2011-02-15  46        0
#> 47  Freiburg 2011-02-16  47        0
#> 48  Freiburg 2011-02-17  48        0
#> 49  Freiburg 2011-02-18  49        0
#> 50  Freiburg 2011-02-19  50        0
#> 51  Freiburg 2011-02-20  51        0
#> 52  Freiburg 2011-02-21  52        0
#> 53  Freiburg 2011-02-22  53        0
#> 54  Freiburg 2011-02-23  54        0
#> 55  Freiburg 2011-02-24  55        0
#> 56  Freiburg 2011-02-25  56        0
#> 57  Freiburg 2011-02-26  57        0
#> 58  Freiburg 2011-02-27  58        0
#> 59  Freiburg 2011-02-28  59        0
#> 60  Freiburg 2011-03-01  60        0
#> 61  Freiburg 2011-03-02  61        0
#> 62  Freiburg 2011-03-03  62        0
#> 63  Freiburg 2011-03-04  63        0
#> 64  Freiburg 2011-03-05  64        0
#> 65  Freiburg 2011-03-06  65        0
#> 66  Freiburg 2011-03-07  66        0
#> 67  Freiburg 2011-03-08  67        0
#> 68  Freiburg 2011-03-09  68        0
#> 69  Freiburg 2011-03-10  69        0
#> 70  Freiburg 2011-03-11  70        0
#> 71  Freiburg 2011-03-12  71        0
#> 72  Freiburg 2011-03-13  72        0
#> 73  Freiburg 2011-03-14  73        0
#> 74  Freiburg 2011-03-15  74   243890
#> 75  Freiburg 2011-03-16  75  1435894
#> 76  Freiburg 2011-03-17  76        0
#> 77  Freiburg 2011-03-18  77        0
#> 78  Freiburg 2011-03-19  78        0
#> 79  Freiburg 2011-03-20  79        0
#> 80  Freiburg 2011-03-21  80        0
#> 81  Freiburg 2011-03-22  81        0
#> 82  Freiburg 2011-03-23  82        0
#> 83  Freiburg 2011-03-24  83  4598833
#> 84  Freiburg 2011-03-25  84  7554480
#> 85  Freiburg 2011-03-26  85 17234414
#> 86  Freiburg 2011-03-27  86  2679499
#> 87  Freiburg 2011-03-28  87        0
#> 88  Freiburg 2011-03-29  88        0
#> 89  Freiburg 2011-03-30  89 10012380
#> 90  Freiburg 2011-03-31  90        0
#> 91  Freiburg 2011-04-01  91        0
#> 92  Freiburg 2011-04-02  92 27778474
#> 93  Freiburg 2011-04-03  93 52206087
#> 94  Freiburg 2011-04-04  94 44261028
#> 95  Freiburg 2011-04-05  95        0
#> 96  Freiburg 2011-04-06  96 15750998
#> 97  Freiburg 2011-04-07  97 44214397
#> 98  Freiburg 2011-04-08  98 56413085
#> 99  Freiburg 2011-04-09  99 30647044
#> 100 Freiburg 2011-04-10 100 29912804
#> 101 Freiburg 2011-04-11 101 32790743
#> 102 Freiburg 2011-04-12 102 40066879
#> 103 Freiburg 2011-04-13 103        0
#> 104 Freiburg 2011-04-14 104        0
#> 105 Freiburg 2011-04-15 105        0
#> 106 Freiburg 2011-04-16 106        0
#> 107 Freiburg 2011-04-17 107  1468471
#> 108 Freiburg 2011-04-18 108 10506187
#> 109 Freiburg 2011-04-19 109 26328501
#> 110 Freiburg 2011-04-20 110 30251230
#> 111 Freiburg 2011-04-21 111 33405532
#> 112 Freiburg 2011-04-22 112 34479096
#> 113 Freiburg 2011-04-23 113 33091243
#> 114 Freiburg 2011-04-24 114 33237896
#> 115 Freiburg 2011-04-25 115 23966054
#> 116 Freiburg 2011-04-26 116 24569835
#> 117 Freiburg 2011-04-27 117 13316430
#> 118 Freiburg 2011-04-28 118  6156406
#> 119 Freiburg 2011-04-29 119  8863069
#> 120 Freiburg 2011-04-30 120        0
#> 121 Freiburg 2011-05-01 121  6191441
#> 122 Freiburg 2011-05-02 122 16819751
#> 123 Freiburg 2011-05-03 123  9641741
#> 124 Freiburg 2011-05-04 124        0
#> 125 Freiburg 2011-05-05 125    76185
#> 126 Freiburg 2011-05-06 126  9703019
#> 127 Freiburg 2011-05-07 127 20330908
#> 128 Freiburg 2011-05-08 128 21929250
#> 129 Freiburg 2011-05-09 129 20006666
#> 130 Freiburg 2011-05-10 130 18630387
#> 131 Freiburg 2011-05-11 131 18295767
#> 132 Freiburg 2011-05-12 132 14957915
#> 133 Freiburg 2011-05-13 133  8093248
#> 134 Freiburg 2011-05-14 134  8000151
#> 135 Freiburg 2011-05-15 135  6184208
#> 136 Freiburg 2011-05-16 136        0
#> 137 Freiburg 2011-05-17 137  1141854
#> 138 Freiburg 2011-05-18 138  7107456
#> 139 Freiburg 2011-05-19 139 10787559
#> 140 Freiburg 2011-05-20 140 10384892
#> 141 Freiburg 2011-05-21 141  8783232
#> 142 Freiburg 2011-05-22 142  7899722
#> 143 Freiburg 2011-05-23 143  7675668
#> 144 Freiburg 2011-05-24 144  6123346
#> 145 Freiburg 2011-05-25 145  5771209
#> 146 Freiburg 2011-05-26 146  5180935
#> 147 Freiburg 2011-05-27 147  5505961
#> 148 Freiburg 2011-05-28 148  1308215
#> 149 Freiburg 2011-05-29 149  2288602
#> 150 Freiburg 2011-05-30 150  3982116
#> 151 Freiburg 2011-05-31 151  4336519
#> 152 Freiburg 2011-06-01 152  2686723
#> 153 Freiburg 2011-06-02 153        0
#> 154 Freiburg 2011-06-03 154  2746378
#> 155 Freiburg 2011-06-04 155  3295125
#> 156 Freiburg 2011-06-05 156  2702340
#> 157 Freiburg 2011-06-06 157  2678522
#> 158 Freiburg 2011-06-07 158  1909427
#> 159 Freiburg 2011-06-08 159  1965342
#> 160 Freiburg 2011-06-09 160        0
#> 161 Freiburg 2011-06-10 161   538488
#> 162 Freiburg 2011-06-11 162  1052945
#> 163 Freiburg 2011-06-12 163   753127
#> 164 Freiburg 2011-06-13 164  1228570
#> 165 Freiburg 2011-06-14 165  1067004
#> 166 Freiburg 2011-06-15 166  1223776
#> 167 Freiburg 2011-06-16 167  1231388
#> 168 Freiburg 2011-06-17 168  1142171
#> 169 Freiburg 2011-06-18 169   838856
#> 170 Freiburg 2011-06-19 170   372756
#> 171 Freiburg 2011-06-20 171   394747
#> 172 Freiburg 2011-06-21 172   662316
#> 173 Freiburg 2011-06-22 173   866034
#> 174 Freiburg 2011-06-23 174   671016
#> 175 Freiburg 2011-06-24 175   415990
#> 176 Freiburg 2011-06-25 176   276052
#> 177 Freiburg 2011-06-26 177   357459
#> 178 Freiburg 2011-06-27 178   481570
#> 179 Freiburg 2011-06-28 179   341792
#> 180 Freiburg 2011-06-29 180   292649
#> 181 Freiburg 2011-06-30 181   297250
#> 182 Freiburg 2011-07-01 182   185462
#> 183 Freiburg 2011-07-02 183   108387
#> 184 Freiburg 2011-07-03 184   117418
#> 185 Freiburg 2011-07-04 185   176587
#> 186 Freiburg 2011-07-05 186   225093
#> 187 Freiburg 2011-07-06 187   232441
#> 188 Freiburg 2011-07-07 188   212804
#> 189 Freiburg 2011-07-08 189   164730
#> 190 Freiburg 2011-07-09 190   137748
#> 191 Freiburg 2011-07-10 191   149356
#> 192 Freiburg 2011-07-11 192   101680
#> 193 Freiburg 2011-07-12 193   116777
#> 194 Freiburg 2011-07-13 194   120463
#> 195 Freiburg 2011-07-14 195    76982
#> 196 Freiburg 2011-07-15 196    39064
#> 197 Freiburg 2011-07-16 197    47145
#> 198 Freiburg 2011-07-17 198    71333
#> 199 Freiburg 2011-07-18 199    58022
#> 200 Freiburg 2011-07-19 200    33377
#> 201 Freiburg 2011-07-20 201    33493
#> 202 Freiburg 2011-07-21 202    29626
#> 203 Freiburg 2011-07-22 203    33031
#> 204 Freiburg 2011-07-23 204    34149
#> 205 Freiburg 2011-07-24 205    21709
#> 206 Freiburg 2011-07-25 206      793
#> 207 Freiburg 2011-07-26 207    29153
#> 208 Freiburg 2011-07-27 208    32138
#> 209 Freiburg 2011-07-28 209    30434
#> 210 Freiburg 2011-07-29 210    24923
#> 211 Freiburg 2011-07-30 211    24040
#> 212 Freiburg 2011-07-31 212    16713
#> 213 Freiburg 2011-08-01 213    12613
#> 214 Freiburg 2011-08-02 214    19528
#> 215 Freiburg 2011-08-03 215    23983
#> 216 Freiburg 2011-08-04 216    23057
#> 217 Freiburg 2011-08-05 217    19738
#> 218 Freiburg 2011-08-06 218    15477
#> 219 Freiburg 2011-08-07 219    13579
#> 220 Freiburg 2011-08-08 220     9146
#> 221 Freiburg 2011-08-09 221     6761
#> 222 Freiburg 2011-08-10 222     4890
#> 223 Freiburg 2011-08-11 223     4051
#> 224 Freiburg 2011-08-12 224     8655
#> 225 Freiburg 2011-08-13 225     8788
#> 226 Freiburg 2011-08-14 226     7612
#> 227 Freiburg 2011-08-15 227     7711
#> 228 Freiburg 2011-08-16 228     4979
#> 229 Freiburg 2011-08-17 229     5037
#> 230 Freiburg 2011-08-18 230     5946
#> 231 Freiburg 2011-08-19 231     5436
#> 232 Freiburg 2011-08-20 232     4414
#> 233 Freiburg 2011-08-21 233     4236
#> 234 Freiburg 2011-08-22 234     3461
#> 235 Freiburg 2011-08-23 235     1623
#> 236 Freiburg 2011-08-24 236     2019
#> 237 Freiburg 2011-08-25 237     2798
#> 238 Freiburg 2011-08-26 238     2217
#> 239 Freiburg 2011-08-27 239     1904
#> 240 Freiburg 2011-08-28 240      364
#> 241 Freiburg 2011-08-29 241     1223
#> 242 Freiburg 2011-08-30 242      975
#> 243 Freiburg 2011-08-31 243      804
#> 244 Freiburg 2011-09-01 244     1082
#> 245 Freiburg 2011-09-02 245     1303
#> 246 Freiburg 2011-09-03 246     1165
#> 247 Freiburg 2011-09-04 247     1134
#> 248 Freiburg 2011-09-05 248      766
#> 249 Freiburg 2011-09-06 249      518
#> 250 Freiburg 2011-09-07 250      594
#>  [ reached 'max' / getOption("max.print") -- omitted 115 rows ]
# }