Create a data frame of day lengths

Generate a data frame of day lengths for given latitudes. The package geosphere is required to use this function.

Usage

create_daylength_df(lat, dates, .quiet = FALSE)

Arguments

lat

Data frame with the fields station and lat. Defines the latitude for the respective stations.

dates

Dates that should be processed.

.quiet

If TRUE, messages are suppressed.

Value

A data frame with the columns date, station and daylength.

See also

Examples

# dates of interest
date_start <- as.Date('2020-01-01')
date_end <- as.Date('2020-12-31')

# calculate day length
dl <- create_daylength_df(barrks_data('station_coords'),
                          seq(date_start, date_end, by = 'day'),
                          .quiet = TRUE)

# print day lengths of station 'Freiburg'
dl[dl$station == 'Freiburg',]
#>            date  station daylength
#> 1    2020-01-01 Freiburg  8.435464
#> 5    2020-01-02 Freiburg  8.451082
#> 9    2020-01-03 Freiburg  8.468145
#> 13   2020-01-04 Freiburg  8.486635
#> 17   2020-01-05 Freiburg  8.506531
#> 21   2020-01-06 Freiburg  8.527813
#> 25   2020-01-07 Freiburg  8.550457
#> 29   2020-01-08 Freiburg  8.574441
#> 33   2020-01-09 Freiburg  8.599739
#> 37   2020-01-10 Freiburg  8.626326
#> 41   2020-01-11 Freiburg  8.654176
#> 45   2020-01-12 Freiburg  8.683262
#> 49   2020-01-13 Freiburg  8.713556
#> 53   2020-01-14 Freiburg  8.745028
#> 57   2020-01-15 Freiburg  8.777651
#> 61   2020-01-16 Freiburg  8.811395
#> 65   2020-01-17 Freiburg  8.846230
#> 69   2020-01-18 Freiburg  8.882127
#> 73   2020-01-19 Freiburg  8.919054
#> 77   2020-01-20 Freiburg  8.956983
#> 81   2020-01-21 Freiburg  8.995882
#> 85   2020-01-22 Freiburg  9.035722
#> 89   2020-01-23 Freiburg  9.076472
#> 93   2020-01-24 Freiburg  9.118103
#> 97   2020-01-25 Freiburg  9.160584
#> 101  2020-01-26 Freiburg  9.203886
#> 105  2020-01-27 Freiburg  9.247981
#> 109  2020-01-28 Freiburg  9.292838
#> 113  2020-01-29 Freiburg  9.338431
#> 117  2020-01-30 Freiburg  9.384731
#> 121  2020-01-31 Freiburg  9.431711
#> 125  2020-02-01 Freiburg  9.479344
#> 129  2020-02-02 Freiburg  9.527603
#> 133  2020-02-03 Freiburg  9.576462
#> 137  2020-02-04 Freiburg  9.625898
#> 141  2020-02-05 Freiburg  9.675884
#> 145  2020-02-06 Freiburg  9.726397
#> 149  2020-02-07 Freiburg  9.777414
#> 153  2020-02-08 Freiburg  9.828912
#> 157  2020-02-09 Freiburg  9.880869
#> 161  2020-02-10 Freiburg  9.933263
#> 165  2020-02-11 Freiburg  9.986074
#> 169  2020-02-12 Freiburg 10.039281
#> 173  2020-02-13 Freiburg 10.092866
#> 177  2020-02-14 Freiburg 10.146808
#> 181  2020-02-15 Freiburg 10.201091
#> 185  2020-02-16 Freiburg 10.255696
#> 189  2020-02-17 Freiburg 10.310606
#> 193  2020-02-18 Freiburg 10.365806
#> 197  2020-02-19 Freiburg 10.421279
#> 201  2020-02-20 Freiburg 10.477009
#> 205  2020-02-21 Freiburg 10.532983
#> 209  2020-02-22 Freiburg 10.589187
#> 213  2020-02-23 Freiburg 10.645606
#> 217  2020-02-24 Freiburg 10.702227
#> 221  2020-02-25 Freiburg 10.759039
#> 225  2020-02-26 Freiburg 10.816029
#> 229  2020-02-27 Freiburg 10.873185
#> 233  2020-02-28 Freiburg 10.930496
#> 237  2020-02-29 Freiburg 10.987952
#> 241  2020-03-01 Freiburg 11.045542
#> 245  2020-03-02 Freiburg 11.103256
#> 249  2020-03-03 Freiburg 11.161085
#> 253  2020-03-04 Freiburg 11.219020
#> 257  2020-03-05 Freiburg 11.277051
#> 261  2020-03-06 Freiburg 11.335171
#> 265  2020-03-07 Freiburg 11.393370
#> 269  2020-03-08 Freiburg 11.451642
#> 273  2020-03-09 Freiburg 11.509978
#> 277  2020-03-10 Freiburg 11.568370
#> 281  2020-03-11 Freiburg 11.626813
#> 285  2020-03-12 Freiburg 11.685298
#> 289  2020-03-13 Freiburg 11.743819
#> 293  2020-03-14 Freiburg 11.802369
#> 297  2020-03-15 Freiburg 11.860942
#> 301  2020-03-16 Freiburg 11.919531
#> 305  2020-03-17 Freiburg 11.978131
#> 309  2020-03-18 Freiburg 12.036734
#> 313  2020-03-19 Freiburg 12.095335
#> 317  2020-03-20 Freiburg 12.153929
#> 321  2020-03-21 Freiburg 12.212508
#> 325  2020-03-22 Freiburg 12.271068
#> 329  2020-03-23 Freiburg 12.329602
#> 333  2020-03-24 Freiburg 12.388104
#> 337  2020-03-25 Freiburg 12.446568
#> 341  2020-03-26 Freiburg 12.504989
#> 345  2020-03-27 Freiburg 12.563360
#> 349  2020-03-28 Freiburg 12.621675
#> 353  2020-03-29 Freiburg 12.679929
#> 357  2020-03-30 Freiburg 12.738114
#> 361  2020-03-31 Freiburg 12.796224
#> 365  2020-04-01 Freiburg 12.854254
#> 369  2020-04-02 Freiburg 12.912195
#> 373  2020-04-03 Freiburg 12.970041
#> 377  2020-04-04 Freiburg 13.027786
#> 381  2020-04-05 Freiburg 13.085421
#> 385  2020-04-06 Freiburg 13.142940
#> 389  2020-04-07 Freiburg 13.200334
#> 393  2020-04-08 Freiburg 13.257595
#> 397  2020-04-09 Freiburg 13.314716
#> 401  2020-04-10 Freiburg 13.371687
#> 405  2020-04-11 Freiburg 13.428500
#> 409  2020-04-12 Freiburg 13.485146
#> 413  2020-04-13 Freiburg 13.541614
#> 417  2020-04-14 Freiburg 13.597896
#> 421  2020-04-15 Freiburg 13.653981
#> 425  2020-04-16 Freiburg 13.709859
#> 429  2020-04-17 Freiburg 13.765518
#> 433  2020-04-18 Freiburg 13.820947
#> 437  2020-04-19 Freiburg 13.876134
#> 441  2020-04-20 Freiburg 13.931067
#> 445  2020-04-21 Freiburg 13.985734
#> 449  2020-04-22 Freiburg 14.040121
#> 453  2020-04-23 Freiburg 14.094215
#> 457  2020-04-24 Freiburg 14.148001
#> 461  2020-04-25 Freiburg 14.201466
#> 465  2020-04-26 Freiburg 14.254594
#> 469  2020-04-27 Freiburg 14.307369
#> 473  2020-04-28 Freiburg 14.359776
#> 477  2020-04-29 Freiburg 14.411798
#> 481  2020-04-30 Freiburg 14.463418
#> 485  2020-05-01 Freiburg 14.514619
#> 489  2020-05-02 Freiburg 14.565383
#> 493  2020-05-03 Freiburg 14.615690
#> 497  2020-05-04 Freiburg 14.665522
#> 501  2020-05-05 Freiburg 14.714860
#> 505  2020-05-06 Freiburg 14.763682
#> 509  2020-05-07 Freiburg 14.811970
#> 513  2020-05-08 Freiburg 14.859701
#> 517  2020-05-09 Freiburg 14.906854
#> 521  2020-05-10 Freiburg 14.953407
#> 525  2020-05-11 Freiburg 14.999337
#> 529  2020-05-12 Freiburg 15.044622
#> 533  2020-05-13 Freiburg 15.089239
#> 537  2020-05-14 Freiburg 15.133163
#> 541  2020-05-15 Freiburg 15.176371
#> 545  2020-05-16 Freiburg 15.218839
#> 549  2020-05-17 Freiburg 15.260541
#> 553  2020-05-18 Freiburg 15.301452
#> 557  2020-05-19 Freiburg 15.341549
#> 561  2020-05-20 Freiburg 15.380804
#> 565  2020-05-21 Freiburg 15.419193
#> 569  2020-05-22 Freiburg 15.456691
#> 573  2020-05-23 Freiburg 15.493271
#> 577  2020-05-24 Freiburg 15.528908
#> 581  2020-05-25 Freiburg 15.563576
#> 585  2020-05-26 Freiburg 15.597250
#> 589  2020-05-27 Freiburg 15.629904
#> 593  2020-05-28 Freiburg 15.661514
#> 597  2020-05-29 Freiburg 15.692055
#> 601  2020-05-30 Freiburg 15.721502
#> 605  2020-05-31 Freiburg 15.749830
#> 609  2020-06-01 Freiburg 15.777017
#> 613  2020-06-02 Freiburg 15.803040
#> 617  2020-06-03 Freiburg 15.827876
#> 621  2020-06-04 Freiburg 15.851503
#> 625  2020-06-05 Freiburg 15.873900
#> 629  2020-06-06 Freiburg 15.895048
#> 633  2020-06-07 Freiburg 15.914926
#> 637  2020-06-08 Freiburg 15.933517
#> 641  2020-06-09 Freiburg 15.950802
#> 645  2020-06-10 Freiburg 15.966767
#> 649  2020-06-11 Freiburg 15.981395
#> 653  2020-06-12 Freiburg 15.994672
#> 657  2020-06-13 Freiburg 16.006587
#> 661  2020-06-14 Freiburg 16.017127
#> 665  2020-06-15 Freiburg 16.026281
#> 669  2020-06-16 Freiburg 16.034042
#> 673  2020-06-17 Freiburg 16.040402
#> 677  2020-06-18 Freiburg 16.045354
#> 681  2020-06-19 Freiburg 16.048895
#> 685  2020-06-20 Freiburg 16.051020
#> 689  2020-06-21 Freiburg 16.051728
#> 693  2020-06-22 Freiburg 16.051020
#> 697  2020-06-23 Freiburg 16.048896
#> 701  2020-06-24 Freiburg 16.045359
#> 705  2020-06-25 Freiburg 16.040413
#> 709  2020-06-26 Freiburg 16.034064
#> 713  2020-06-27 Freiburg 16.026318
#> 717  2020-06-28 Freiburg 16.017186
#> 721  2020-06-29 Freiburg 16.006675
#> 725  2020-06-30 Freiburg 15.994798
#> 729  2020-07-01 Freiburg 15.981566
#> 733  2020-07-02 Freiburg 15.966994
#> 737  2020-07-03 Freiburg 15.951095
#> 741  2020-07-04 Freiburg 15.933888
#> 745  2020-07-05 Freiburg 15.915387
#> 749  2020-07-06 Freiburg 15.895612
#> 753  2020-07-07 Freiburg 15.874581
#> 757  2020-07-08 Freiburg 15.852314
#> 761  2020-07-09 Freiburg 15.828833
#> 765  2020-07-10 Freiburg 15.804158
#> 769  2020-07-11 Freiburg 15.778312
#> 773  2020-07-12 Freiburg 15.751318
#> 777  2020-07-13 Freiburg 15.723199
#> 781  2020-07-14 Freiburg 15.693979
#> 785  2020-07-15 Freiburg 15.663682
#> 789  2020-07-16 Freiburg 15.632333
#> 793  2020-07-17 Freiburg 15.599957
#> 797  2020-07-18 Freiburg 15.566580
#> 801  2020-07-19 Freiburg 15.532226
#> 805  2020-07-20 Freiburg 15.496922
#> 809  2020-07-21 Freiburg 15.460693
#> 813  2020-07-22 Freiburg 15.423564
#> 817  2020-07-23 Freiburg 15.385561
#> 821  2020-07-24 Freiburg 15.346710
#> 825  2020-07-25 Freiburg 15.307035
#> 829  2020-07-26 Freiburg 15.266563
#> 833  2020-07-27 Freiburg 15.225318
#> 837  2020-07-28 Freiburg 15.183325
#> 841  2020-07-29 Freiburg 15.140608
#> 845  2020-07-30 Freiburg 15.097192
#> 849  2020-07-31 Freiburg 15.053099
#> 853  2020-08-01 Freiburg 15.008355
#> 857  2020-08-02 Freiburg 14.962980
#> 861  2020-08-03 Freiburg 14.916999
#> 865  2020-08-04 Freiburg 14.870433
#> 869  2020-08-05 Freiburg 14.823303
#> 873  2020-08-06 Freiburg 14.775632
#> 877  2020-08-07 Freiburg 14.727439
#> 881  2020-08-08 Freiburg 14.678745
#> 885  2020-08-09 Freiburg 14.629570
#> 889  2020-08-10 Freiburg 14.579933
#> 893  2020-08-11 Freiburg 14.529851
#> 897  2020-08-12 Freiburg 14.479345
#> 901  2020-08-13 Freiburg 14.428431
#> 905  2020-08-14 Freiburg 14.377126
#> 909  2020-08-15 Freiburg 14.325447
#> 913  2020-08-16 Freiburg 14.273410
#> 917  2020-08-17 Freiburg 14.221031
#> 921  2020-08-18 Freiburg 14.168324
#> 925  2020-08-19 Freiburg 14.115305
#> 929  2020-08-20 Freiburg 14.061987
#> 933  2020-08-21 Freiburg 14.008385
#> 937  2020-08-22 Freiburg 13.954511
#> 941  2020-08-23 Freiburg 13.900378
#> 945  2020-08-24 Freiburg 13.845998
#> 949  2020-08-25 Freiburg 13.791383
#> 953  2020-08-26 Freiburg 13.736545
#> 957  2020-08-27 Freiburg 13.681494
#> 961  2020-08-28 Freiburg 13.626242
#> 965  2020-08-29 Freiburg 13.570798
#> 969  2020-08-30 Freiburg 13.515172
#> 973  2020-08-31 Freiburg 13.459373
#> 977  2020-09-01 Freiburg 13.403412
#> 981  2020-09-02 Freiburg 13.347296
#> 985  2020-09-03 Freiburg 13.291034
#> 989  2020-09-04 Freiburg 13.234634
#> 993  2020-09-05 Freiburg 13.178105
#> 997  2020-09-06 Freiburg 13.121453
#> 1001 2020-09-07 Freiburg 13.064687
#> 1005 2020-09-08 Freiburg 13.007813
#> 1009 2020-09-09 Freiburg 12.950838
#> 1013 2020-09-10 Freiburg 12.893769
#> 1017 2020-09-11 Freiburg 12.836613
#> 1021 2020-09-12 Freiburg 12.779376
#> 1025 2020-09-13 Freiburg 12.722063
#> 1029 2020-09-14 Freiburg 12.664682
#> 1033 2020-09-15 Freiburg 12.607237
#> 1037 2020-09-16 Freiburg 12.549735
#> 1041 2020-09-17 Freiburg 12.492182
#> 1045 2020-09-18 Freiburg 12.434582
#> 1049 2020-09-19 Freiburg 12.376942
#> 1053 2020-09-20 Freiburg 12.319267
#> 1057 2020-09-21 Freiburg 12.261562
#> 1061 2020-09-22 Freiburg 12.203833
#> 1065 2020-09-23 Freiburg 12.146085
#> 1069 2020-09-24 Freiburg 12.088324
#> 1073 2020-09-25 Freiburg 12.030555
#> 1077 2020-09-26 Freiburg 11.972784
#> 1081 2020-09-27 Freiburg 11.915015
#> 1085 2020-09-28 Freiburg 11.857256
#> 1089 2020-09-29 Freiburg 11.799511
#> 1093 2020-09-30 Freiburg 11.741786
#> 1097 2020-10-01 Freiburg 11.684088
#> 1101 2020-10-02 Freiburg 11.626423
#> 1105 2020-10-03 Freiburg 11.568796
#> 1109 2020-10-04 Freiburg 11.511215
#> 1113 2020-10-05 Freiburg 11.453686
#> 1117 2020-10-06 Freiburg 11.396215
#> 1121 2020-10-07 Freiburg 11.338811
#> 1125 2020-10-08 Freiburg 11.281480
#> 1129 2020-10-09 Freiburg 11.224230
#> 1133 2020-10-10 Freiburg 11.167069
#> 1137 2020-10-11 Freiburg 11.110006
#> 1141 2020-10-12 Freiburg 11.053048
#> 1145 2020-10-13 Freiburg 10.996205
#> 1149 2020-10-14 Freiburg 10.939487
#> 1153 2020-10-15 Freiburg 10.882902
#> 1157 2020-10-16 Freiburg 10.826461
#> 1161 2020-10-17 Freiburg 10.770174
#> 1165 2020-10-18 Freiburg 10.714052
#> 1169 2020-10-19 Freiburg 10.658108
#> 1173 2020-10-20 Freiburg 10.602351
#> 1177 2020-10-21 Freiburg 10.546795
#> 1181 2020-10-22 Freiburg 10.491453
#> 1185 2020-10-23 Freiburg 10.436338
#> 1189 2020-10-24 Freiburg 10.381463
#> 1193 2020-10-25 Freiburg 10.326844
#> 1197 2020-10-26 Freiburg 10.272495
#> 1201 2020-10-27 Freiburg 10.218432
#> 1205 2020-10-28 Freiburg 10.164671
#> 1209 2020-10-29 Freiburg 10.111228
#> 1213 2020-10-30 Freiburg 10.058121
#> 1217 2020-10-31 Freiburg 10.005369
#> 1221 2020-11-01 Freiburg  9.952989
#> 1225 2020-11-02 Freiburg  9.901001
#> 1229 2020-11-03 Freiburg  9.849425
#> 1233 2020-11-04 Freiburg  9.798281
#> 1237 2020-11-05 Freiburg  9.747591
#> 1241 2020-11-06 Freiburg  9.697376
#> 1245 2020-11-07 Freiburg  9.647659
#> 1249 2020-11-08 Freiburg  9.598463
#> 1253 2020-11-09 Freiburg  9.549812
#> 1257 2020-11-10 Freiburg  9.501730
#> 1261 2020-11-11 Freiburg  9.454243
#> 1265 2020-11-12 Freiburg  9.407375
#> 1269 2020-11-13 Freiburg  9.361152
#> 1273 2020-11-14 Freiburg  9.315603
#> 1277 2020-11-15 Freiburg  9.270753
#> 1281 2020-11-16 Freiburg  9.226631
#> 1285 2020-11-17 Freiburg  9.183264
#> 1289 2020-11-18 Freiburg  9.140682
#> 1293 2020-11-19 Freiburg  9.098913
#> 1297 2020-11-20 Freiburg  9.057987
#> 1301 2020-11-21 Freiburg  9.017932
#> 1305 2020-11-22 Freiburg  8.978780
#> 1309 2020-11-23 Freiburg  8.940560
#> 1313 2020-11-24 Freiburg  8.903303
#> 1317 2020-11-25 Freiburg  8.867038
#> 1321 2020-11-26 Freiburg  8.831796
#> 1325 2020-11-27 Freiburg  8.797606
#> 1329 2020-11-28 Freiburg  8.764500
#>  [ reached 'max' / getOption("max.print") -- omitted 33 rows ]