Create a data frame of sunrises and sunsets

Generate a data frame that specifies sunrises and sunsets for different coordinates and dates. The package suncalc is required to use this function.

Usage

create_suntimes_df(coords, dates, tz = Sys.timezone(), .quiet = FALSE)

Arguments

coords

Data frame with the fields station, lat and lon. Defines the latitude and longitude for the respective stations.

dates

Dates that should be processed.

tz

Timezone of the results.

.quiet

If TRUE, messages are suppressed.

Value

A data frame with the columns date, station and sunrise and sunset. The values of sunrise and sunset indicate the respective time in minutes.

See also

Examples

# \donttest{
date_start <- as.Date('2020-01-01')
date_end <- as.Date('2020-12-31')

st <- create_suntimes_df(barrks_data('station_coords'),
                         seq(date_start, date_end, by = 'day'),
                         .quiet = TRUE)

# print results of station 'Freiburg'
st[st$station == 'Freiburg',]
#>           date  station sunrise sunset
#> 1   2020-01-01 Freiburg     499   1006
#> 5   2020-01-02 Freiburg     499   1007
#> 9   2020-01-03 Freiburg     499   1008
#> 13  2020-01-04 Freiburg     499   1009
#> 17  2020-01-05 Freiburg     499   1010
#> 21  2020-01-06 Freiburg     499   1011
#> 25  2020-01-07 Freiburg     499   1012
#> 29  2020-01-08 Freiburg     498   1013
#> 33  2020-01-09 Freiburg     498   1014
#> 37  2020-01-10 Freiburg     498   1015
#> 41  2020-01-11 Freiburg     497   1017
#> 45  2020-01-12 Freiburg     497   1018
#> 49  2020-01-13 Freiburg     496   1019
#> 53  2020-01-14 Freiburg     495   1021
#> 57  2020-01-15 Freiburg     495   1022
#> 61  2020-01-16 Freiburg     494   1023
#> 65  2020-01-17 Freiburg     494   1025
#> 69  2020-01-18 Freiburg     493   1026
#> 73  2020-01-19 Freiburg     492   1028
#> 77  2020-01-20 Freiburg     491   1029
#> 81  2020-01-21 Freiburg     490   1031
#> 85  2020-01-22 Freiburg     489   1032
#> 89  2020-01-23 Freiburg     488   1034
#> 93  2020-01-24 Freiburg     487   1035
#> 97  2020-01-25 Freiburg     486   1037
#> 101 2020-01-26 Freiburg     485   1038
#> 105 2020-01-27 Freiburg     484   1040
#> 109 2020-01-28 Freiburg     483   1041
#> 113 2020-01-29 Freiburg     482   1043
#> 117 2020-01-30 Freiburg     481   1044
#> 121 2020-01-31 Freiburg     480   1046
#> 125 2020-02-01 Freiburg     478   1048
#> 129 2020-02-02 Freiburg     477   1049
#> 133 2020-02-03 Freiburg     476   1051
#> 137 2020-02-04 Freiburg     474   1052
#> 141 2020-02-05 Freiburg     473   1054
#> 145 2020-02-06 Freiburg     471   1056
#> 149 2020-02-07 Freiburg     470   1057
#> 153 2020-02-08 Freiburg     468   1059
#> 157 2020-02-09 Freiburg     467   1060
#> 161 2020-02-10 Freiburg     465   1062
#> 165 2020-02-11 Freiburg     464   1064
#> 169 2020-02-12 Freiburg     462   1065
#> 173 2020-02-13 Freiburg     461   1067
#> 177 2020-02-14 Freiburg     459   1068
#> 181 2020-02-15 Freiburg     457   1070
#> 185 2020-02-16 Freiburg     456   1072
#> 189 2020-02-17 Freiburg     454   1073
#> 193 2020-02-18 Freiburg     452   1075
#> 197 2020-02-19 Freiburg     451   1076
#> 201 2020-02-20 Freiburg     449   1078
#> 205 2020-02-21 Freiburg     447   1080
#> 209 2020-02-22 Freiburg     445   1081
#> 213 2020-02-23 Freiburg     443   1083
#> 217 2020-02-24 Freiburg     442   1084
#> 221 2020-02-25 Freiburg     440   1086
#> 225 2020-02-26 Freiburg     438   1087
#> 229 2020-02-27 Freiburg     436   1089
#> 233 2020-02-28 Freiburg     434   1091
#> 237 2020-02-29 Freiburg     432   1092
#> 241 2020-03-01 Freiburg     430   1094
#> 245 2020-03-02 Freiburg     428   1095
#> 249 2020-03-03 Freiburg     426   1097
#> 253 2020-03-04 Freiburg     424   1098
#> 257 2020-03-05 Freiburg     422   1100
#> 261 2020-03-06 Freiburg     420   1101
#> 265 2020-03-07 Freiburg     418   1103
#> 269 2020-03-08 Freiburg     416   1104
#> 273 2020-03-09 Freiburg     414   1106
#> 277 2020-03-10 Freiburg     412   1107
#> 281 2020-03-11 Freiburg     410   1109
#> 285 2020-03-12 Freiburg     408   1110
#> 289 2020-03-13 Freiburg     406   1112
#> 293 2020-03-14 Freiburg     404   1113
#> 297 2020-03-15 Freiburg     402   1115
#> 301 2020-03-16 Freiburg     400   1116
#> 305 2020-03-17 Freiburg     398   1117
#> 309 2020-03-18 Freiburg     396   1119
#> 313 2020-03-19 Freiburg     394   1120
#> 317 2020-03-20 Freiburg     392   1122
#> 321 2020-03-21 Freiburg     390   1123
#> 325 2020-03-22 Freiburg     388   1125
#> 329 2020-03-23 Freiburg     386   1126
#> 333 2020-03-24 Freiburg     384   1128
#> 337 2020-03-25 Freiburg     382   1129
#> 341 2020-03-26 Freiburg     379   1130
#> 345 2020-03-27 Freiburg     377   1132
#> 349 2020-03-28 Freiburg     375   1133
#> 353 2020-03-29 Freiburg     433   1195
#> 357 2020-03-30 Freiburg     431   1196
#> 361 2020-03-31 Freiburg     429   1198
#> 365 2020-04-01 Freiburg     427   1199
#> 369 2020-04-02 Freiburg     425   1200
#> 373 2020-04-03 Freiburg     423   1202
#> 377 2020-04-04 Freiburg     421   1203
#> 381 2020-04-05 Freiburg     419   1205
#> 385 2020-04-06 Freiburg     417   1206
#> 389 2020-04-07 Freiburg     415   1208
#> 393 2020-04-08 Freiburg     413   1209
#> 397 2020-04-09 Freiburg     411   1211
#> 401 2020-04-10 Freiburg     409   1212
#> 405 2020-04-11 Freiburg     407   1213
#> 409 2020-04-12 Freiburg     405   1215
#> 413 2020-04-13 Freiburg     403   1216
#> 417 2020-04-14 Freiburg     401   1218
#> 421 2020-04-15 Freiburg     399   1219
#> 425 2020-04-16 Freiburg     397   1221
#> 429 2020-04-17 Freiburg     395   1222
#> 433 2020-04-18 Freiburg     394   1223
#> 437 2020-04-19 Freiburg     392   1225
#> 441 2020-04-20 Freiburg     390   1226
#> 445 2020-04-21 Freiburg     388   1228
#> 449 2020-04-22 Freiburg     386   1229
#> 453 2020-04-23 Freiburg     384   1231
#> 457 2020-04-24 Freiburg     383   1232
#> 461 2020-04-25 Freiburg     381   1234
#> 465 2020-04-26 Freiburg     379   1235
#> 469 2020-04-27 Freiburg     377   1236
#> 473 2020-04-28 Freiburg     376   1238
#> 477 2020-04-29 Freiburg     374   1239
#> 481 2020-04-30 Freiburg     372   1241
#> 485 2020-05-01 Freiburg     371   1242
#> 489 2020-05-02 Freiburg     369   1244
#> 493 2020-05-03 Freiburg     368   1245
#> 497 2020-05-04 Freiburg     366   1246
#> 501 2020-05-05 Freiburg     364   1248
#> 505 2020-05-06 Freiburg     363   1249
#> 509 2020-05-07 Freiburg     361   1251
#> 513 2020-05-08 Freiburg     360   1252
#> 517 2020-05-09 Freiburg     358   1253
#> 521 2020-05-10 Freiburg     357   1255
#> 525 2020-05-11 Freiburg     356   1256
#> 529 2020-05-12 Freiburg     354   1257
#> 533 2020-05-13 Freiburg     353   1259
#> 537 2020-05-14 Freiburg     352   1260
#> 541 2020-05-15 Freiburg     350   1261
#> 545 2020-05-16 Freiburg     349   1263
#> 549 2020-05-17 Freiburg     348   1264
#> 553 2020-05-18 Freiburg     347   1265
#> 557 2020-05-19 Freiburg     346   1267
#> 561 2020-05-20 Freiburg     345   1268
#> 565 2020-05-21 Freiburg     343   1269
#> 569 2020-05-22 Freiburg     342   1270
#> 573 2020-05-23 Freiburg     341   1271
#> 577 2020-05-24 Freiburg     340   1273
#> 581 2020-05-25 Freiburg     340   1274
#> 585 2020-05-26 Freiburg     339   1275
#> 589 2020-05-27 Freiburg     338   1276
#> 593 2020-05-28 Freiburg     337   1277
#> 597 2020-05-29 Freiburg     336   1278
#> 601 2020-05-30 Freiburg     335   1279
#> 605 2020-05-31 Freiburg     335   1280
#> 609 2020-06-01 Freiburg     334   1281
#> 613 2020-06-02 Freiburg     333   1282
#> 617 2020-06-03 Freiburg     333   1283
#> 621 2020-06-04 Freiburg     332   1284
#> 625 2020-06-05 Freiburg     332   1284
#> 629 2020-06-06 Freiburg     331   1285
#> 633 2020-06-07 Freiburg     331   1286
#> 637 2020-06-08 Freiburg     331   1287
#> 641 2020-06-09 Freiburg     330   1287
#> 645 2020-06-10 Freiburg     330   1288
#> 649 2020-06-11 Freiburg     330   1289
#> 653 2020-06-12 Freiburg     330   1289
#> 657 2020-06-13 Freiburg     329   1290
#> 661 2020-06-14 Freiburg     329   1290
#> 665 2020-06-15 Freiburg     329   1291
#> 669 2020-06-16 Freiburg     329   1291
#> 673 2020-06-17 Freiburg     329   1292
#> 677 2020-06-18 Freiburg     329   1292
#> 681 2020-06-19 Freiburg     329   1292
#> 685 2020-06-20 Freiburg     330   1293
#> 689 2020-06-21 Freiburg     330   1293
#> 693 2020-06-22 Freiburg     330   1293
#> 697 2020-06-23 Freiburg     330   1293
#> 701 2020-06-24 Freiburg     331   1293
#> 705 2020-06-25 Freiburg     331   1293
#> 709 2020-06-26 Freiburg     331   1293
#> 713 2020-06-27 Freiburg     332   1293
#> 717 2020-06-28 Freiburg     332   1293
#> 721 2020-06-29 Freiburg     333   1293
#> 725 2020-06-30 Freiburg     333   1293
#> 729 2020-07-01 Freiburg     334   1293
#> 733 2020-07-02 Freiburg     335   1292
#> 737 2020-07-03 Freiburg     335   1292
#> 741 2020-07-04 Freiburg     336   1292
#> 745 2020-07-05 Freiburg     337   1291
#> 749 2020-07-06 Freiburg     337   1291
#> 753 2020-07-07 Freiburg     338   1290
#> 757 2020-07-08 Freiburg     339   1290
#> 761 2020-07-09 Freiburg     340   1289
#> 765 2020-07-10 Freiburg     341   1289
#> 769 2020-07-11 Freiburg     342   1288
#> 773 2020-07-12 Freiburg     343   1287
#> 777 2020-07-13 Freiburg     344   1286
#> 781 2020-07-14 Freiburg     345   1286
#> 785 2020-07-15 Freiburg     346   1285
#> 789 2020-07-16 Freiburg     347   1284
#> 793 2020-07-17 Freiburg     348   1283
#> 797 2020-07-18 Freiburg     349   1282
#> 801 2020-07-19 Freiburg     350   1281
#> 805 2020-07-20 Freiburg     351   1280
#> 809 2020-07-21 Freiburg     352   1279
#> 813 2020-07-22 Freiburg     353   1278
#> 817 2020-07-23 Freiburg     355   1277
#> 821 2020-07-24 Freiburg     356   1276
#> 825 2020-07-25 Freiburg     357   1275
#> 829 2020-07-26 Freiburg     358   1274
#> 833 2020-07-27 Freiburg     359   1272
#> 837 2020-07-28 Freiburg     361   1271
#> 841 2020-07-29 Freiburg     362   1270
#> 845 2020-07-30 Freiburg     363   1268
#> 849 2020-07-31 Freiburg     365   1267
#> 853 2020-08-01 Freiburg     366   1266
#> 857 2020-08-02 Freiburg     367   1264
#> 861 2020-08-03 Freiburg     368   1263
#> 865 2020-08-04 Freiburg     370   1261
#> 869 2020-08-05 Freiburg     371   1260
#> 873 2020-08-06 Freiburg     372   1258
#> 877 2020-08-07 Freiburg     374   1257
#> 881 2020-08-08 Freiburg     375   1255
#> 885 2020-08-09 Freiburg     376   1253
#> 889 2020-08-10 Freiburg     378   1252
#> 893 2020-08-11 Freiburg     379   1250
#> 897 2020-08-12 Freiburg     381   1249
#> 901 2020-08-13 Freiburg     382   1247
#> 905 2020-08-14 Freiburg     383   1245
#> 909 2020-08-15 Freiburg     385   1243
#> 913 2020-08-16 Freiburg     386   1242
#> 917 2020-08-17 Freiburg     387   1240
#> 921 2020-08-18 Freiburg     389   1238
#> 925 2020-08-19 Freiburg     390   1236
#> 929 2020-08-20 Freiburg     391   1234
#> 933 2020-08-21 Freiburg     393   1233
#> 937 2020-08-22 Freiburg     394   1231
#> 941 2020-08-23 Freiburg     396   1229
#> 945 2020-08-24 Freiburg     397   1227
#> 949 2020-08-25 Freiburg     398   1225
#> 953 2020-08-26 Freiburg     400   1223
#> 957 2020-08-27 Freiburg     401   1221
#> 961 2020-08-28 Freiburg     402   1219
#> 965 2020-08-29 Freiburg     404   1217
#> 969 2020-08-30 Freiburg     405   1215
#> 973 2020-08-31 Freiburg     407   1213
#> 977 2020-09-01 Freiburg     408   1211
#> 981 2020-09-02 Freiburg     409   1209
#> 985 2020-09-03 Freiburg     411   1207
#> 989 2020-09-04 Freiburg     412   1205
#> 993 2020-09-05 Freiburg     413   1203
#> 997 2020-09-06 Freiburg     415   1201
#>  [ reached 'max' / getOption("max.print") -- omitted 116 rows ]
# }