Get (preprocessed) input data

The function returns a list that contains the input data of the phenology as well as some intermediate results that are needed as preprocessed inputs for the model. The result can be used as input for phenology() to avoid redundant calculations.

Usage

get_input_data(pheno)

Arguments

pheno

A phenology (see phenology())

Value

A list of SpatRasters.

Examples

# \donttest{
# setup phenology
p <- phenology('phenips-clim', barrks_data(), .setup_only = TRUE, .quiet = TRUE)

# get the (preprocessed) input data
inputs <- get_input_data(p)

# print the names to show which input data is available
names(inputs)
#>  [1] "daylength"    "rad"          "sunrise"      "sunset"       "tmax"        
#>  [6] "tmean"        "tmin"         "fly"          "dd_onset"     "dd_onset_alt"
#> [11] "btmean"       "btmax"        "teff"        
# }