Use BSO

This page describes the usage of BSO with phenology(). The model-specific inputs are listed and its basic functionality is explained. BSO was published by Jakoby et al. (2019) and parametrized for Ips typographus in Switzerland. Note that the onset and the development submodel do not support the usage of a storage (except for some precalculations).

Arguments

...

See phenology() for a detailled description of the function.

tmin, tmax

Daily minimum/maximum air temperatures in °C.

sunrise, sunset

Time of sunrise/sunset in minutes from midnight. Can be created with create_suntimes_rsts() or create_suntimes_df().

n

number of individuals to simulate.

max_generations

maximum number of generations to calculate.

Value

The function returns a BSO phenology. Look here to find out how it can be analysed.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply BSO and which additional parameters are available.

bso_phenology("bso", ..., tmin, tmax, sunrise, sunset,
              n = 1e+09, max_generations = 4)

# calculate submodels separately
bso_phenology("bso", ..., .submodels = 'onset',
              tmin, tmax, sunrise, sunset, n = 1e+09)
bso_phenology("bso", ..., .submodels = 'diapause', tmin, tmax)
bso_phenology("bso", ..., .submodels = 'development',
              .onset, .diapause = NULL, .mortality = NULL,
              tmin, tmax, sunrise, sunset,
              max_generations = 4)

Functioning of the BSO

In the following, the basic functioning of BSO is explained.

Onset: The onset of swarming will start when the degree days of the mean temperature reach a specific threshold and regeneration feeding of the individuals has finished (Look at development for details).

Development: The development of single individuals is simulated. The simulation of each individual is realized by passing a multitude of slots that are grouped in stages. The hourly probability for an individual to enter the next slot depends on the current stage and the phloem temperature. The hourly temperature is derived from the minimum and maximum temperatures using a sine interpolition. The hourly phloem temperature is calculated using Newton's Law of Cooling (see Trân et al. 2007 ).

Diapause: Specific photoperiod-related dates define when the diapause is initiated at the earliest and at the latest. In between these dates, the diapause is initiated when the mean temperature falls below a specific threshold.

Mortality: BSO does not have a mortality submodel implemented.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

References

Jakoby O, Lischke H, Wermelinger B (2019). “Climate change alters elevational phenology patterns of the European spruce bark beetle (Ips typographus).” Global Change Biology, 25(12), 4048-4063. doi:10.1111/gcb.14766 .

Trân JK, Ylioja T, Billings RF, Régnière J, Ayres MP (2007). “Impact of minimum winter temperatures on the population dynamics of Dendroctonus frontalis.” Ecological Applications, 17(3), 882-899. doi:10.1890/06-0512 .

See also