Use the Jönnson model

This page describes the usage of the the Jönsson model with phenology(). The model specific inputs are listed and its basic functionality is explained. The Jönsson model was published by Jönsson et al. (2011) and parametrized for Ips typographus in southern Sweden.

Arguments

...

See phenology() for a detailled description of the function.

tmean, tmax

Daily mean/maximum temperatures in °C.

daylength

Length of the day in hours. Can be created with create_daylength_rst() or create_daylength_rst().

mode

Can be 'fast' (default) or 'slow'. Determines if the lower ('fast') or upper ('slow') limit for the development of genertion should be used.

Value

The function returns a phenology. Look here to find out how it can be analysed.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply the Jönsson model and which additional parameters are available.

phenology("joensson", ..., tmean, tmax, daylength, mode = 'fast')

# calculate submodels separately
phenology("joensson", ..., .submodels = 'onset', tmax)
phenology("joensson", ..., .submodels = 'diapause', tmax, daylength)
phenology("joensson", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmean, mode = 'fast')

Functioning

In the following, the basic functioning of the Jönsson model is explained.

Onset: The onset of swarming is triggerd when the degree days of the maximum temperature reach a specific threshold and the maximum temperature exceeds the minimum flight temperature. The onset of infestation is triggered seven days later to account for a pre-oviposition period.

Development: The development progresses proportional to the degree days of the mean temperature. To account for varying sun exposures, two different thermal thresholds are defined that reflect the lower and the upper limit of development. A generations starts swarming when it has finished its development and the maximum temperature exceeds the minimum flight temperature. Seven days later, the development of a new generation starts.

Diapause: The diapause is initiated when the daylength falls below a threshold. It is recommended to adjust the daylength threshold when applying elsewhere (e.g. values from literature). Jönsson et al. (2011) proposes a model to calculate the daylength threshold based on long-term climate data.

Mortality: The Jönsson model does not have a mortality submodel implemented.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

References

Jönsson AM, Harding S, Krokene P, Lange H, Åke Lindelöw A, Økland B, Ravn HP, Schroeder LM (2011). “Modelling the potential impact of global warming on Ips typographus voltinism and reproductive diapause.” Climatic Change, 109, 695-718. doi:10.1007/s10584-011-0038-4 .

See also