Use PHENIPS

This page describes the usage of PHENIPS with phenology(). The model specific inputs are listed and its basic functionality is explained. PHENIPS was published by Baier et al. (2007) and parametrized at the Kalkalpen National Park in Austria for Ips typographus.

Arguments

...

See phenology() for a detailled description of the function.

tmean, tmax

Daily mean/maximum temperatures in °C.

rad

Daily radiation in W * h / m^2.

daylength

Length of the day in hours. Can be created with create_daylength_rst() or create_daylength_df().

exposure

Specifies the sun exposure. Can be 'sunny' (default) or 'shaded'.

sister_broods

Set FALSE if sister broods should not be calculated.

Value

The function returns a phenology. Look here to find out how it can be analysed.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply PHENIPS and which additional parameters are available.

phenology("phenips", ..., tmean, tmax, rad, daylength,
          exposure = 'sunny', sister_broods = TRUE)

# calculate submodels separately
phenology("phenips", ..., .submodels = 'onset', tmax)
phenology("phenips", ..., .submodels = 'diapause', daylength)
phenology("phenips", ..., .submodels = 'mortality', tmax)
phenology("phenips", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmean, tmax, rad,
          exposure = 'sunny', sister_broods = TRUE)

Functioning

In the following, the basic functioning of PHENIPS is explained.

Onset: The onset is triggered when the degree days of the maximum temperature reach a specific threshold and the maximum temperature exceeds the minimum flight temperature.

Development: The beetles develop according to a slightly modified version of the optimum curve described by Wermelinger and Seifert (1998) depending on the bark temperature. The bark temperature is modeled based on mean and maximum temperature, global radiation and sun exposure. A new generation will emerge when the last generation is fully developed and the maximum temperature exceeds the minimum flight temperature.

Diapause: The diapause is initiated when the daylength falls below a threshold.

Mortality: White stages (egg to pupa) die on a fixed date.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

References

Baier P, Pennerstorfer J, Schopf A (2007). “PHENIPS—A comprehensive phenology model of Ips typographus (L.)(Col., Scolytinae) as a tool for hazard rating of bark beetle infestation.” Forest Ecology and Management, 249(3), 171–186. doi:10.1016/j.foreco.2007.05.020 .

Wermelinger B, Seifert M (1998). “Analysis of the temperature dependent development of the spruce bark beetle Ips typographus (L) (Col., Scolytidae).” Journal of Applied Entomology, 122(1-5), 185-191. doi:10.1111/j.1439-0418.1998.tb01482.x .

See also