Use PHENIPS-Clim

This page describes the usage of PHENIPS-Clim with phenology(). The model specific inputs are listed and its basic functionality is explained. PHENIPS-Clim is not published yet. This manual will be updated when a publication is available. It was parametrized for Ips typographus in southern Germany.

Arguments

...

See phenology() for a detailled description of the function. See phenology() for details.

tmin, tmean, tmax

Daily minimum/mean/maximum temperatures in °C. tmin is optional. If available it will be used to calculate the temperature amplitude. If not, (tmax - tmean) * 2 will be used as amplitude.

rad

Daily radiation in W * h / m^2.

daylength

Length of the day in hours. Can be created with create_daylength_rst() or create_daylength_rst().

sister_broods

Set FALSE to disable the calculation of sister broods.

scenario

Choose a scenario to use a suitable combination of parameters for specific situations. The scenario defines a default value for each value that can be overwritten by specifying a value for the respective parameter. The following scenarios are available:

mean: list(exposure = 'sunny', onset_mode = 0.5, diapause_mode = 'photoperiodic', oviposition_mode = 0.5)

max: list(exposure = 'sunny', onset_mode = 0.1, diapause_mode = 'thermal', oviposition_mode = 0.1)

exposure

Specifies the sun exposure. Can be 'sunny'(default) or 'shaded'.

onset_mode

Share of beetles that are already infesting trees necessary to trigger the onset. Must be 0.1, 0.5 or 0.9 if not customized.

oviposition_mode

Share of beetles that should have finished oviposition to trigger the beginning of the development. Must be 0.1, 0.5 or 0.9 if not customized.

diapause_mode

Determines how the diapause is initiated. Can be one of the following options:

'photoperiodic': The diapause is initiated when the daylength falls below a specific threshold.

'thermal': The diapause is initiated by a logistic model that depends on the daylength and the maximum temperature.

Share of beetles that already stopped reproducing necessary to trigger the diapause. Must be thermal or 'photoperiodic' if not customized. If 'photoperiodic' is chosen, the diapause is controlled by a daylength threshold (see parameter daylength_dia here).

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply PHENIPS-Clim and which additional parameters are available.

phenology("phenips-clim", ..., tmin, tmean, tmax, rad, daylength,
          sister_broods = TRUE, scenario = 'max', exposure = NULL,
          onset_mode = NULL, oviposition_mode = NULL, diapause_mode = NULL)

# calculate submodels separately
phenology("phenips-clim", ..., .submodels = 'onset', tmax, scenario = 'max', onset_mode = NULL)
phenology("phenips-clim", ..., .submodels = 'diapause', tmax, daylength, scenario = 'max', diapause_mode = NULL)
phenology("phenips-clim", ..., .submodels = 'mortality', tmin)
phenology("phenips-clim", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmin, tmean, tmax, rad, sister_broods = TRUE,
          scenario = 'max', exposure = NULL, oviposition_mode = NULL)

Functioning

In the following, the basic functioning of PHENIPS-Clim is explained.

Onset: A base onset is triggered by a logistic model that relates to the maximum temperature and the respective degree days. Beginning from the base onset, a specific level of degree days (depending on the share of individuals that should be accounted for) and maximum air temperature must be reached to trigger the actual onset.

Development: While the bark temperature and the emergence of new generations are determined according to PHENIPS, the calculation of the beetles' development rates is refined. Rather than implying a constant development within a day, temperature fluctuations are incorporated by taking the daily temperature amplitude into account. Additionally, the first part of development represents the pre-oviposition period and will not appear in the resulting output.

Diapause: The diapause can be initiated due to the photoperiod according to PHENIPS or by a logistic model that depends on the daylength and the maximum temperature and accounts for beetles that reproduce even on shorter days if the temperatures are favorable. In the second case, PHENIPS-Clim detects a reproductive arrest, due to adverse abiotic parameters, and not an actual diapause as the output can be adjusted, if conditions improve and allow for further reproduction later in the season.

Mortality: White stages (egg to pupa) die when the minimum temperature falls below a specific threshold.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

See also