Customize PHENIPS-Clim

In barrks, model() is used to customize a model. Here, the parameters are described that can be used to customize PHENIPS-Clim. The model is currently unpublished. This manual will be updated as soon as a publication is available. Look here to find out how to apply the model.

Usage

model("phenips-clim",

      # ==== onset ====

      dd_onset_start_date = '03-01',
      dd_onset_base = 12,
      onset_func = \(tmax, dd_tmax) 0.564071 * tmax + 0.006434 * dd_tmax - 12.37046 > 0,
      onset_add_dd = c('0.1' = 0, '0.5' = 90, '0.9' = 190),

      # ==== development ====

      model_end_date = '12-31',
      tfly = 16.5,

      dd_total_dev = 557,

      dev_start = 0.15,
      dev_end = 1,
      dev_sister_brood = 0.5,
      dev_oviposition = c('0' = 0,
                          '0.1' = 0.1,
                          '0.5' = 0.15,
                          '0.9' = 0.26),

      dev_mortal_min = 0,
      dev_mortal_max = 0.6,

      topt = 30.4,

      func_btmean = function(tmean, rad) { -0.173 + 0.0008518 * rad + 1.054 * tmean},
      func_btmax = function(tmax, rad) { 1.656 + 0.002955 * rad + 0.534 * tmax + 0.01884 * tmax ^ 2 },
      func_btdiff = function(tmax) { (-310.667 + 9.603 * tmax) / 24 },

      dev_rates = phenips_clim_get_dev_rates(),

      # ==== diapause ====

      first_diapause_date = '08-12',
      diapause_thermal_func = \(daylength, tmax) 0.8619156 * daylength + 0.5081128 * tmax - 23.63691 > 0,
      daylength_dia = 14.5,

      # ==== mortality ====

      tlethal = -5

)

Arguments

dd_onset_start_date

The date, when the degree days start to sum up ('MM-DD').

dd_onset_base

Base temperature to calculate degree days to trigger the onset.

onset_func

Function with the SpatRasters tmax (maximum temperature) and dd_tmax (degree days of maximum temperature) as parameters. The function should return TRUE when the base onset is triggered. See onset_add_dd for the actual onset of infestation.

onset_add_dd

Vector of options to calculate the actual onset of infestation. The vector should be named after the share of beetles that already started breeding when the onset is triggered (choose an option via phenology(..., onset_mode = [option]) when applying the model). The values specify the degree days that are required starting at the

model_end_date

Date when the model ends.

tfly

Minimum temperature that beetles need to fly.

dd_total_dev

Degree days that are required for a generation to fully develop

dev_start, dev_end

Share in total development when the egg development starts and the juvenile beetle's development ends respectively. Usable if the development below/above these thresholds should account for mating, oviposition etc.

dev_sister_brood

Share in the total development, when a sister brood will be established.

dev_oviposition

Named numeric vector of shares in the total development when the oviposition is finished. The vector should be named after the share of beetles that should be taken into account (choose an option via phenology(..., oviposition_mode = [option]) when applying the model).

dev_mortal_min, dev_mortal_max

Minimum/maximum share in the total development of white stages (egg, larva, pupa). During these stages, the beetles could die caused by a mortality event.

topt

Temperature for optimal development.

func_btmean, func_btmax, func_btdiff

Functions to calculate the bark temperatures (see Baier et al. 2007 , equations A.3 to A.5). Each parameter will be passed as SpatRaster:

tmean: mean air temperature

tmax: maximum air temperature

rad: radiation

btmax: maximum bark temperature

dev_rates

Data frame that specifies the development rates per day depending on the mean temperature and the temperature amplitude. Column names are the mean temperatures and row names the temperature amplitudes both with one decimal place. base onset (see onset_func) to trigger the actual onset.

first_diapause_date

Date before which an initiation of the diapause is impossible ('MM-DD').

diapause_thermal_func

Function to calculate the initiation of the diapause if the model was applied using phenology(..., diapause_mode = 'thermal'). The diapause will be initiated the last time when the function returns TRUE.

daylength_dia

When the daylength falls below this threshold, diapause will be initiated if the model was applied using phenology(..., diapause_mode = 'photoperiodic').

tlethal

Temperature threshold below which white stages will die.

References

Baier P, Pennerstorfer J, Schopf A (2007). “PHENIPS—A comprehensive phenology model of Ips typographus (L.)(Col., Scolytinae) as a tool for hazard rating of bark beetle infestation.” Forest Ecology and Management, 249(3), 171--186. doi:10.1016/j.foreco.2007.05.020 .

See also