Translate BSO generated phenology

A BSO generated phenology cannot be analysed in the same way as other phenology objects. To be able to use the functions that are available for phenology objects returned by phenology(), the BSO generated phenology should be translated.

Usage

bso_translate_phenology(pheno, threshold = 0.1, .quiet = FALSE)

Arguments

pheno

A BSO phenology (see bso_phenology())

threshold

Share of individuals that must have reached a specific development in the BSO phenology to account for them in the corresponding standard phenology.

.quiet

If TRUE, messages are suppressed.

Value

Returns a standard phenology as a list. Look here to find out how a phenology can be analysed. It is not recommended to access the list elements directly.

Examples

# \donttest{
# This may take a few minutes...

# calculate and translate BSO phenology
p <- bso_phenology('bso', barrks_data('stations'), .quiet = TRUE)
pt <- bso_translate_phenology(p, .quiet = TRUE)

# print the generations data frame of station 'Freiburg'
df <- get_generations_df(pt, 'Freiburg')
df
#>      station       date doy generation
#> 1   Freiburg 2011-01-01   1        0.0
#> 2   Freiburg 2011-01-02   2        0.0
#> 3   Freiburg 2011-01-03   3        0.0
#> 4   Freiburg 2011-01-04   4        0.0
#> 5   Freiburg 2011-01-05   5        0.0
#> 6   Freiburg 2011-01-06   6        0.0
#> 7   Freiburg 2011-01-07   7        0.0
#> 8   Freiburg 2011-01-08   8        0.0
#> 9   Freiburg 2011-01-09   9        0.0
#> 10  Freiburg 2011-01-10  10        0.0
#> 11  Freiburg 2011-01-11  11        0.0
#> 12  Freiburg 2011-01-12  12        0.0
#> 13  Freiburg 2011-01-13  13        0.0
#> 14  Freiburg 2011-01-14  14        0.0
#> 15  Freiburg 2011-01-15  15        0.0
#> 16  Freiburg 2011-01-16  16        0.0
#> 17  Freiburg 2011-01-17  17        0.0
#> 18  Freiburg 2011-01-18  18        0.0
#> 19  Freiburg 2011-01-19  19        0.0
#> 20  Freiburg 2011-01-20  20        0.0
#> 21  Freiburg 2011-01-21  21        0.0
#> 22  Freiburg 2011-01-22  22        0.0
#> 23  Freiburg 2011-01-23  23        0.0
#> 24  Freiburg 2011-01-24  24        0.0
#> 25  Freiburg 2011-01-25  25        0.0
#> 26  Freiburg 2011-01-26  26        0.0
#> 27  Freiburg 2011-01-27  27        0.0
#> 28  Freiburg 2011-01-28  28        0.0
#> 29  Freiburg 2011-01-29  29        0.0
#> 30  Freiburg 2011-01-30  30        0.0
#> 31  Freiburg 2011-01-31  31        0.0
#> 32  Freiburg 2011-02-01  32        0.0
#> 33  Freiburg 2011-02-02  33        0.0
#> 34  Freiburg 2011-02-03  34        0.0
#> 35  Freiburg 2011-02-04  35        0.0
#> 36  Freiburg 2011-02-05  36        0.0
#> 37  Freiburg 2011-02-06  37        0.0
#> 38  Freiburg 2011-02-07  38        0.0
#> 39  Freiburg 2011-02-08  39        0.0
#> 40  Freiburg 2011-02-09  40        0.0
#> 41  Freiburg 2011-02-10  41        0.0
#> 42  Freiburg 2011-02-11  42        0.0
#> 43  Freiburg 2011-02-12  43        0.0
#> 44  Freiburg 2011-02-13  44        0.0
#> 45  Freiburg 2011-02-14  45        0.0
#> 46  Freiburg 2011-02-15  46        0.0
#> 47  Freiburg 2011-02-16  47        0.0
#> 48  Freiburg 2011-02-17  48        0.0
#> 49  Freiburg 2011-02-18  49        0.0
#> 50  Freiburg 2011-02-19  50        0.0
#> 51  Freiburg 2011-02-20  51        0.0
#> 52  Freiburg 2011-02-21  52        0.0
#> 53  Freiburg 2011-02-22  53        0.0
#> 54  Freiburg 2011-02-23  54        0.0
#> 55  Freiburg 2011-02-24  55        0.0
#> 56  Freiburg 2011-02-25  56        0.0
#> 57  Freiburg 2011-02-26  57        0.0
#> 58  Freiburg 2011-02-27  58        0.0
#> 59  Freiburg 2011-02-28  59        0.0
#> 60  Freiburg 2011-03-01  60        0.0
#> 61  Freiburg 2011-03-02  61        0.0
#> 62  Freiburg 2011-03-03  62        0.0
#> 63  Freiburg 2011-03-04  63        0.0
#> 64  Freiburg 2011-03-05  64        0.0
#> 65  Freiburg 2011-03-06  65        0.0
#> 66  Freiburg 2011-03-07  66        0.0
#> 67  Freiburg 2011-03-08  67        0.0
#> 68  Freiburg 2011-03-09  68        0.0
#> 69  Freiburg 2011-03-10  69        0.0
#> 70  Freiburg 2011-03-11  70        0.0
#> 71  Freiburg 2011-03-12  71        0.0
#> 72  Freiburg 2011-03-13  72        0.0
#> 73  Freiburg 2011-03-14  73        0.0
#> 74  Freiburg 2011-03-15  74        0.0
#> 75  Freiburg 2011-03-16  75        0.0
#> 76  Freiburg 2011-03-17  76        0.0
#> 77  Freiburg 2011-03-18  77        0.0
#> 78  Freiburg 2011-03-19  78        0.0
#> 79  Freiburg 2011-03-20  79        0.0
#> 80  Freiburg 2011-03-21  80        0.0
#> 81  Freiburg 2011-03-22  81        0.0
#> 82  Freiburg 2011-03-23  82        0.0
#> 83  Freiburg 2011-03-24  83        0.0
#> 84  Freiburg 2011-03-25  84        0.0
#> 85  Freiburg 2011-03-26  85        0.0
#> 86  Freiburg 2011-03-27  86        0.0
#> 87  Freiburg 2011-03-28  87        0.0
#> 88  Freiburg 2011-03-29  88        0.0
#> 89  Freiburg 2011-03-30  89        0.0
#> 90  Freiburg 2011-03-31  90        0.0
#> 91  Freiburg 2011-04-01  91        0.0
#> 92  Freiburg 2011-04-02  92        0.0
#> 93  Freiburg 2011-04-03  93        1.0
#> 94  Freiburg 2011-04-04  94        1.0
#> 95  Freiburg 2011-04-05  95        1.0
#> 96  Freiburg 2011-04-06  96        1.0
#> 97  Freiburg 2011-04-07  97        1.0
#> 98  Freiburg 2011-04-08  98        1.0
#> 99  Freiburg 2011-04-09  99        1.0
#> 100 Freiburg 2011-04-10 100        1.0
#> 101 Freiburg 2011-04-11 101        1.0
#> 102 Freiburg 2011-04-12 102        1.5
#> 103 Freiburg 2011-04-13 103        1.5
#> 104 Freiburg 2011-04-14 104        1.5
#> 105 Freiburg 2011-04-15 105        1.5
#> 106 Freiburg 2011-04-16 106        1.5
#> 107 Freiburg 2011-04-17 107        1.5
#> 108 Freiburg 2011-04-18 108        1.5
#> 109 Freiburg 2011-04-19 109        1.5
#> 110 Freiburg 2011-04-20 110        1.5
#> 111 Freiburg 2011-04-21 111        1.5
#> 112 Freiburg 2011-04-22 112        1.5
#> 113 Freiburg 2011-04-23 113        1.5
#> 114 Freiburg 2011-04-24 114        1.5
#> 115 Freiburg 2011-04-25 115        1.5
#> 116 Freiburg 2011-04-26 116        1.5
#> 117 Freiburg 2011-04-27 117        1.5
#> 118 Freiburg 2011-04-28 118        1.5
#> 119 Freiburg 2011-04-29 119        1.5
#> 120 Freiburg 2011-04-30 120        1.5
#> 121 Freiburg 2011-05-01 121        1.5
#> 122 Freiburg 2011-05-02 122        1.5
#> 123 Freiburg 2011-05-03 123        1.5
#> 124 Freiburg 2011-05-04 124        1.5
#> 125 Freiburg 2011-05-05 125        1.5
#> 126 Freiburg 2011-05-06 126        1.5
#> 127 Freiburg 2011-05-07 127        1.5
#> 128 Freiburg 2011-05-08 128        1.5
#> 129 Freiburg 2011-05-09 129        1.5
#> 130 Freiburg 2011-05-10 130        1.5
#> 131 Freiburg 2011-05-11 131        1.5
#> 132 Freiburg 2011-05-12 132        1.5
#> 133 Freiburg 2011-05-13 133        1.5
#> 134 Freiburg 2011-05-14 134        1.5
#> 135 Freiburg 2011-05-15 135        1.5
#> 136 Freiburg 2011-05-16 136        1.5
#> 137 Freiburg 2011-05-17 137        1.5
#> 138 Freiburg 2011-05-18 138        1.5
#> 139 Freiburg 2011-05-19 139        1.5
#> 140 Freiburg 2011-05-20 140        1.5
#> 141 Freiburg 2011-05-21 141        1.5
#> 142 Freiburg 2011-05-22 142        1.5
#> 143 Freiburg 2011-05-23 143        1.5
#> 144 Freiburg 2011-05-24 144        1.5
#> 145 Freiburg 2011-05-25 145        1.5
#> 146 Freiburg 2011-05-26 146        1.5
#> 147 Freiburg 2011-05-27 147        1.5
#> 148 Freiburg 2011-05-28 148        1.5
#> 149 Freiburg 2011-05-29 149        1.5
#> 150 Freiburg 2011-05-30 150        1.5
#> 151 Freiburg 2011-05-31 151        1.5
#> 152 Freiburg 2011-06-01 152        1.5
#> 153 Freiburg 2011-06-02 153        1.5
#> 154 Freiburg 2011-06-03 154        1.5
#> 155 Freiburg 2011-06-04 155        1.5
#> 156 Freiburg 2011-06-05 156        1.5
#> 157 Freiburg 2011-06-06 157        1.5
#> 158 Freiburg 2011-06-07 158        1.5
#> 159 Freiburg 2011-06-08 159        1.5
#> 160 Freiburg 2011-06-09 160        1.5
#> 161 Freiburg 2011-06-10 161        1.5
#> 162 Freiburg 2011-06-11 162        1.5
#> 163 Freiburg 2011-06-12 163        1.5
#> 164 Freiburg 2011-06-13 164        1.5
#> 165 Freiburg 2011-06-14 165        1.5
#> 166 Freiburg 2011-06-15 166        1.5
#> 167 Freiburg 2011-06-16 167        1.5
#> 168 Freiburg 2011-06-17 168        1.5
#> 169 Freiburg 2011-06-18 169        1.5
#> 170 Freiburg 2011-06-19 170        1.5
#> 171 Freiburg 2011-06-20 171        1.5
#> 172 Freiburg 2011-06-21 172        1.5
#> 173 Freiburg 2011-06-22 173        1.5
#> 174 Freiburg 2011-06-23 174        1.5
#> 175 Freiburg 2011-06-24 175        1.5
#> 176 Freiburg 2011-06-25 176        1.5
#> 177 Freiburg 2011-06-26 177        1.5
#> 178 Freiburg 2011-06-27 178        1.5
#> 179 Freiburg 2011-06-28 179        2.0
#> 180 Freiburg 2011-06-29 180        2.0
#> 181 Freiburg 2011-06-30 181        2.0
#> 182 Freiburg 2011-07-01 182        2.0
#> 183 Freiburg 2011-07-02 183        2.0
#> 184 Freiburg 2011-07-03 184        2.0
#> 185 Freiburg 2011-07-04 185        2.0
#> 186 Freiburg 2011-07-05 186        2.5
#> 187 Freiburg 2011-07-06 187        2.5
#> 188 Freiburg 2011-07-07 188        2.5
#> 189 Freiburg 2011-07-08 189        2.5
#> 190 Freiburg 2011-07-09 190        2.5
#> 191 Freiburg 2011-07-10 191        2.5
#> 192 Freiburg 2011-07-11 192        2.5
#> 193 Freiburg 2011-07-12 193        2.5
#> 194 Freiburg 2011-07-13 194        2.5
#> 195 Freiburg 2011-07-14 195        2.5
#> 196 Freiburg 2011-07-15 196        2.5
#> 197 Freiburg 2011-07-16 197        2.5
#> 198 Freiburg 2011-07-17 198        2.5
#> 199 Freiburg 2011-07-18 199        2.5
#> 200 Freiburg 2011-07-19 200        2.5
#> 201 Freiburg 2011-07-20 201        2.5
#> 202 Freiburg 2011-07-21 202        2.5
#> 203 Freiburg 2011-07-22 203        2.5
#> 204 Freiburg 2011-07-23 204        2.5
#> 205 Freiburg 2011-07-24 205        2.5
#> 206 Freiburg 2011-07-25 206        2.5
#> 207 Freiburg 2011-07-26 207        2.5
#> 208 Freiburg 2011-07-27 208        2.5
#> 209 Freiburg 2011-07-28 209        2.5
#> 210 Freiburg 2011-07-29 210        2.5
#> 211 Freiburg 2011-07-30 211        2.5
#> 212 Freiburg 2011-07-31 212        2.5
#> 213 Freiburg 2011-08-01 213        2.5
#> 214 Freiburg 2011-08-02 214        2.5
#> 215 Freiburg 2011-08-03 215        2.5
#> 216 Freiburg 2011-08-04 216        2.5
#> 217 Freiburg 2011-08-05 217        2.5
#> 218 Freiburg 2011-08-06 218        2.5
#> 219 Freiburg 2011-08-07 219        2.5
#> 220 Freiburg 2011-08-08 220        2.5
#> 221 Freiburg 2011-08-09 221        2.5
#> 222 Freiburg 2011-08-10 222        2.5
#> 223 Freiburg 2011-08-11 223        2.5
#> 224 Freiburg 2011-08-12 224        2.5
#> 225 Freiburg 2011-08-13 225        2.5
#> 226 Freiburg 2011-08-14 226        2.5
#> 227 Freiburg 2011-08-15 227        2.5
#> 228 Freiburg 2011-08-16 228        2.5
#> 229 Freiburg 2011-08-17 229        2.5
#> 230 Freiburg 2011-08-18 230        2.5
#> 231 Freiburg 2011-08-19 231        2.5
#> 232 Freiburg 2011-08-20 232        2.5
#> 233 Freiburg 2011-08-21 233        2.5
#> 234 Freiburg 2011-08-22 234        2.5
#> 235 Freiburg 2011-08-23 235        2.5
#> 236 Freiburg 2011-08-24 236        2.5
#> 237 Freiburg 2011-08-25 237        2.5
#> 238 Freiburg 2011-08-26 238        2.5
#> 239 Freiburg 2011-08-27 239        2.5
#> 240 Freiburg 2011-08-28 240        2.5
#> 241 Freiburg 2011-08-29 241        2.5
#> 242 Freiburg 2011-08-30 242        2.5
#> 243 Freiburg 2011-08-31 243        2.5
#> 244 Freiburg 2011-09-01 244        2.5
#> 245 Freiburg 2011-09-02 245        2.5
#> 246 Freiburg 2011-09-03 246        2.5
#> 247 Freiburg 2011-09-04 247        2.5
#> 248 Freiburg 2011-09-05 248        2.5
#> 249 Freiburg 2011-09-06 249        2.5
#> 250 Freiburg 2011-09-07 250        2.5
#>  [ reached 'max' / getOption("max.print") -- omitted 115 rows ]
# }