Use CHAPY

This page describes the usage of CHAPY with phenology(). The model specific inputs are listed and its basic functionality is explained. CHAPY was published by Ogris et al. (2020) and parametrized for Pityogenes chalcographus in Slovenia.

Arguments

tmin, tmean, tmax

Daily minimum/mean/maximum air temperatures in °C. For the development submodel, the parameter that is obligatory depends on mode.

daylength

Length of the day in hours. Can be created with create_daylength_rst() or create_daylength_rst().

mode

Specifies which temperature should be used to calculate the development. Can be min, mean or max.

.submodels, .onset, .diapause, .mortality, ...

See phenology() for a detailled description of the function.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply CHAPY and which additional parameters are available.

phenology("chapy", ..., tmin = NULL, tmean = NULL, tmax, daylength, mode = 'max')

# calculate submodels separately
phenology("chapy", ..., .submodels = 'onset', tmax)
phenology("chapy", ..., .submodels = 'diapause', daylength)
phenology("chapy", ..., .submodels = 'mortality', tmax)
phenology("chapy", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmin = NULL, tmean = NULL, tmax = NULL, mode = 'max')

Functioning

The functioning of CHAPY is identical to RITY but it is has a different parametrization.

References

Ogris N, Ferlan M, Hauptman T, Pavlin R, Kavčič A, Jurc M, de Groot M (2020). “Sensitivity analysis, calibration and validation of a phenology model for Pityogenes chalcographus (CHAPY).” Ecological Modelling, 430, 109137. ISSN 0304-3800, doi:10.1016/j.ecolmodel.2020.109137 .

See also