Customize the Jönsson model

This page describes the parameters that can be used to customize the Jönsson model. The model was developed by Jönsson et al. (2011) . Look here to find out how to apply the model.

Arguments

dd_onset_start_date

The date, when the degree days start to sum up ('MM-DD').

dd_onset_base

Base temperature to calculate degree days to trigger the onset.

dd_onset_threshold

Degree days that are required to trigger the onset of infestation. Additionally, the maximum temperature must exceed tfly.

tfly

Minimum temperature that beetles need to fly.

dd_development_base

Base temperature to calculate degree days for development.

dd_total_dev_lower, dd_total_dev_upper

Lower/upper limit of degree days that are required for a generation to fully develop

dev_start, dev_end

Share in total development when the egg development starts and the juvenile beetle's development ends respectively. Usable if the development below/above these thresholds should account for mating, oviposition etc.

dev_mortal_min, dev_mortal_max

The beetles are considered to be in white stages (egg, larva, pupa) if their development exceeds dev_mortal_min and subceeds dev_mortal_max. During these stages, the beetles could die due to a mortality event. NULL means that no lower/upper threshold is defined.

model_end_date

Date when the model ends (no further development will be modeled).

daylength_dia, tdia_min

When the daylength falls below daylength_dia and the average daily temperature falls below tdia_min, diapause will be initiated. The default value for the critical daylength was set to 19.3 hours according to Schroeder and Dalin (2017) who examined the photoperiodic diapause induction in Sweden. If the model is used for other regions, this value should be adjusted.

Details

In barrks, model() is used to customize a model. The following code illustrates which parameters are available for the Jönsson model and specifies their default values.

model("joensson",

      # ==== onset ====

      dd_onset_start_date = '01-01',
      dd_onset_base = 5,
      dd_onset_threshold = 120,

      # ==== onset + development ====

      tfly = 20,

      # ==== development ====

      dd_development_base = 5,
      dd_total_dev_lower = 625,
      dd_total_dev_upper = 750,
      dev_start = 0,
      dev_end = 1,
      dev_mortal_min = 0,
      dev_mortal_max = 1,

      model_end_date = '12-31',

      # ==== diapause ====

      daylength_dia = 19.3,
      tdia_min = 15
)

References

Jönsson AM, Harding S, Krokene P, Lange H, Åke Lindelöw A, Økland B, Ravn HP, Schroeder LM (2011). “Modelling the potential impact of global warming on Ips typographus voltinism and reproductive diapause.” Climatic Change, 109, 695-718. doi:10.1007/s10584-011-0038-4 .

Schroeder M, Dalin P (2017). “Differences in photoperiod-induced diapause plasticity among different populations of the bark beetle Ips typographus and its predator Thanasimus formicarius.” Agricultural and Forest Entomology, 19(2), 146-153. doi:10.1111/afe.12189 .

See also