Use the Lange model

This page describes the usage of the the Lange model with phenology(). The model specific inputs are listed and its basic functionality is explained. The model was published by Lange et al. (2008) for Ips typographus.

Arguments

tmin, tmean, tmax

Daily minimum/mean/maximum temperatures in °C.

.submodels, .onset, .diapause, .mortality, ...

See phenology() for a detailled description of the function.

Value

The function returns a phenology. Look here to find out how it can be analysed.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply the Lange model and which additional parameters are available.

phenology("lange", ..., tmin, tmean, tmax)

# calculate submodels separately
phenology("lange", ..., .submodels = 'onset', tmean, tmax)
phenology("lange", ..., .submodels = 'mortality', tmin)
phenology("lange", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmean, tmax)

Functioning

In the following, the basic functioning of the Lange model is explained.

Onset: The onset of swarming is triggerd when the degree days of the maximum temperature reach a specific threshold and the maximum temperature exceeds the minimum flight temperature according to Annila (1969) .

Development: The development is calculated using stage-specific temperature sums and thresholds (Wermelinger and Seifert 1998) . A new generation starts its development when the last generation finished its development and the maximum temperature exceeds the minimum flight temperature.

Diapause: The Lange model does not have a diapause submodel implemented.

Mortality: White stages (egg to pupa) die when the minimum temperature falls below a specific threshold.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

References

Annila E (1969). “Influence of temperature upon the development and voltinism of Ips typographus L. (Coleoptera, Scolytidae).” Annales Zoologici Fennici, 6(2), 161–208. http://www.jstor.org/stable/23731366.

Lange H, Økland B, Krokene P (2008). “To be or twice to be? The life cycle development of the spruce bark beetle under climate change.” In Unifying Themes in Complex Systems: Proceedings of the Sixth International Conference on Complex Systems, 251–258. Springer. doi:10.1007/978-3-540-85081-6_32 .

Wermelinger B, Seifert M (1998). “Analysis of the temperature dependent development of the spruce bark beetle Ips typographus (L) (Col., Scolytidae).” Journal of Applied Entomology, 122(1-5), 185-191. doi:10.1111/j.1439-0418.1998.tb01482.x .

See also