Customize the Lange model

This page describes the parameters that can be used to customize Lange. The model was developed by Lange et al. (2008) . Look here to find out how to apply the model.

Arguments

dd_onset_start_date

The date, when the degree days start to sum up ('MM-DD').

dd_onset_base

Base temperature to calculate degree days to trigger the onset.

dd_onset_threshold

Degree days that are required to trigger the onset of infestation. Additionally, the maximum temperature must exceed tfly.

tfly

Minimum temperature that beetles need to fly.

dd_base_stages

Base temperatures to calculate degree days for the different stages in the following order: egg, larva, pupa, juvenile adult.

dd_threshold_stages

Thermal thresholds for the different stages in the following order: egg, larva, pupa, juvenile adult.

model_end_date

Date when the model ends (no further development will be modeled).

first_lethal_date

Date before which no mortality will be modeled.

tlethal

Temperature threshold below which white stages (egg, larva, pupa) will die.

Details

In barrks, model() is used to customize a model. The following code illustrates which parameters are available for the Lange model and specifies their default values.

model("lange",

      # ==== onset ==== #

      dd_onset_start_date = '01-01',
      dd_onset_base = 5,
      dd_onset_threshold = 110,

      # ==== onset + development ====

      tfly = 19.5,

      # ==== development ==== #

      dd_base_stages = c(10.6, 8.2, 9.9, 3.2),
      dd_threshold_stages = c(51.8, 204.4, 57.7, 238.5),

      model_end_date = '12-31',

      # ==== mortality ==== #

      first_lethal_date = '09-01',
      tlethal = 0
)

References

Lange H, Økland B, Krokene P (2008). “To be or twice to be? The life cycle development of the spruce bark beetle under climate change.” In Unifying Themes in Complex Systems: Proceedings of the Sixth International Conference on Complex Systems, 251–258. Springer. doi:10.1007/978-3-540-85081-6_32 .

See also