Use RITY

This page describes the usage of RITY with phenology(). The model specific inputs are listed and its basic functionality is explained. RITY (also called RITY-2) was published by Ogris et al. (2019) and parametrized for Ips typographus in Slovenia.

Arguments

tmin, tmean, tmax

Daily minimum/mean/maximum temperatures in °C. For the development submodel, the parameter that is obligatory depends on mode.

daylength

Length of the day in hours. Can be created with create_daylength_rst() or create_daylength_rst().

mode

Specifies which temperature should be used to calculate the development. Can be min, mean or max.

.submodels, .onset, .diapause, .mortality, ...

See phenology() for a detailled description of the function.

Value

The function returns a phenology. Look here to find out how it can be analysed.

Details

In barrks, phenology() is used to apply a model. The following code illustrates which inputs are required to apply RITY and which additional parameters are available.

phenology("rity", ..., tmin = NULL, tmean = NULL, tmax, daylength, mode = 'max')

# calculate submodels separately
phenology("rity", ..., .submodels = 'onset', tmax)
phenology("rity", ..., .submodels = 'diapause', daylength)
phenology("rity", ..., .submodels = 'mortality', tmax)
phenology("rity", ..., .submodels = 'development',
          .onset, .diapause = NULL, .mortality = NULL,
          tmin = NULL, tmean = NULL, tmax = NULL, mode = 'max')

Functioning

In the following, the basic functioning of RITY is explained.

Onset: See PHENIPS.

Development: Based on PHENIPS with a few modifications:

The optimum curve is calculated according to Wermelinger and Seifert (1998) without simplification.

The minimum, mean or maximum bark temperature can be used to calculate the development. These temperatures depend only on the respective air temperatures.

Diapause: See PHENIPS.

Mortality: See PHENIPS.

Look here to find out how the model parameters affect the actual calculations and which values are used by default.

References

Ogris N, Ferlan M, Hauptman T, Pavlin R, Kavčič A, Jurc M, De Groot M (2019). “RITY--A phenology model of Ips typographus as a tool for optimization of its monitoring.” Ecological Modelling, 410, 108775. doi:10.1016/j.ecolmodel.2019.108775 .

Wermelinger B, Seifert M (1998). “Analysis of the temperature dependent development of the spruce bark beetle Ips typographus (L) (Col., Scolytidae).” Journal of Applied Entomology, 122(1-5), 185-191. doi:10.1111/j.1439-0418.1998.tb01482.x .

See also